misc:stml4xx: Fix WAKE_SENSOR 24-bit extension.
Pushing a change that should have been part of (CR) but Git helpfully
screwed up for me.

Reviewed-on: http://gerrit.mot.com/736264
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ include/linux/motosh.h
@@ -403,7 +403,7 @@ struct motosh_data {
 	wait_queue_head_t quickpeek_wait_queue;
 	atomic_t qp_enabled;
 	bool quickpeek_occurred;
-	unsigned short qw_irq_status;
+	unsigned long qw_irq_status;
 	struct motosh_aod_enabled_vote aod_enabled;
 	bool ignore_wakeable_interrupts;
 	int ignored_interrupts;
